+++ Packages/com.vareiko.foundation/Runtime/Connectivity/INetworkReachabilityProvider.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Vareiko.Foundation.Connectivity
+{
+    public interface INetworkReachabilityProvider
+    {
+        NetworkReachability GetReachability();
+    }
+}