+++ Packages/com.vareiko.foundation/Runtime/Validation/StartupValidationSeverity.cs
@@ -1,0 +1,9 @@
+namespace Vareiko.Foundation.Validation
+{
+    public enum StartupValidationSeverity
+    {
+        Info = 0,
+        Warning = 1,
+        Error = 2
+    }
+}